src/producer/requirements.txt
gevent==21.12.0
greenlet==1.1.2
kafka-python==2.0.2
numpy==1.23.1
pandas==1.4.3
pip==21.1.1
python-dateutil==2.8.2
pytz==2022.1
setuptools==56.0.0
six==1.16.0
websocket-client==1.3.3
zope.event==4.5.0
zope.interface==5.4.0
